saveFibonacci with location as long, not int

+++ src/main/java/maunaloax/models/ChartWindowDressingModel.java
@@ -12,8 +12,8 @@ public interface ChartWindowDressingModel {
     List<DBObject> fetchComments(ObjectId id);
     WriteResult addComment(ObjectId id, String comment);
     WriteResult updateCoord(ObjectId id, DBObject p1, DBObject p2);
-    MongoDBResult saveFibonacci(String ticker, int location, DBObject p1, DBObject p2);
-    List<DBObject> fetchFibonacci(String ticker, int location, Date fromDate, Date toDate);
+    MongoDBResult saveFibonacci(String ticker, long location, DBObject p1, DBObject p2);
+    List<DBObject> fetchFibonacci(String ticker, long location, Date fromDate, Date toDate);
 }
